requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
HTMLTestRunner-rv==1.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
pyparsing==3.1.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0